Refactor : 기존 한 클래스에서 담당하던 전체 절차를 Race & CarManager 로 구분. (입출력&게임순서 담당:Race, 자동차 관련 기능 담당:Car)
