+++ docs/README.md
@@ -1,10 +1,15 @@
 # 🚀 기능 구현 목록
 ## 💼 비즈니스
-### 게임주체자
+### 게임주체자(Race)
 - [x] 이름을 입력해달라는 안내문구를 출력한다.
 - [x] 이름을 입력받는다.
 - [x] 횟수를 입력해달라는 안내문구를 출력한다.
 - [x] 횟수를 입력받는다.
+- [x] 주어진 횟수만큼 한 라운드(Match)를 반복한다.
+   - [x] "자동차 관리자"에게 한 라운드 실행을 요청한다.
+- [x] 우승자를 출력한다.
+
+### 자동차 관리자(CarManager)
 - [x] 입력받은 이름만큼 "자동차"를 생성한다.
 - [x] 각 라운드를 입력받은 차수만큼 반복한다.
    - [x] 전진 여부 결정: 입력받은 자동차 수만큼 반복 실행한다.
@@ -12,7 +17,6 @@
       - [x] "자동차"에게 알려준다.
       - [x] 결과를 "출력담당자"에게 출력하도록 시킨다.
 - [x] 각 자동차들의 거리를 비교하여 최고 거리를 간 자동차를 찾고, 최종 우승자(들)을 찾는다.
-- [x] 우승자를 출력한다.
 
 ### 자동차
 - [x] 이름, 전진상태
@@ -65,7 +69,7 @@
 * * *
 # 예외상황
 - [x] 자동차가 넘겨받은 숫자의 범위가 0 ~ 9 를 넘어서는 경우 IllegalArgumentException예외 발생
-
+- [x] 검증자의 검증이 적절하지 않을 때 IllegarArgumentException 발생시키기.
 
 * * *
 # 체크리스트
