--- src/main/java/racingcar/business/Racing.java
@@ -1,67 +1,0 @@
-package racingcar.business;
-
-import racingcar.ui.InputView;
-import racingcar.ui.Output;
-import racingcar.ui.OutputView;
-
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
-public class Racing {
-
-    InputView inputView = new InputView();
-    OutputView outputView = new OutputView();
-    NumberMaker numberMaker = new NumberMaker();
-
-    List<Car> carList = new ArrayList<>();
-
-    public void doRacing() {
-        outputView.print(Output.GET_NAMES);
-        List<String> names = inputView.getNames();
-
-        outputView.print(Output.GET_GAME_COUNT);
-        Integer gameCount = inputView.getGameCount();
-
-        setCars(names);
-
-        for (int i = 0; i < gameCount; i++) {
-            runRound();
-        }
-
-        String winnerNames = getWinnerNames();
-        outputView.printWinnerNames(winnerNames);
-    }
-
-    private void setCars(List<String> names) {
-        for (String name : names) {
-            carList.add(new Car(name));
-        }
-    }
-
-    private void runRound() {
-        for (Car car : carList) {
-            tryMoving(car);
-        }
-
-        outputView.print(Output.BLANK_LINE);
-    }
-
-    private void tryMoving(Car car) {
-        int randomNumber = numberMaker.getRandomNumber();
-        car.moveCar(randomNumber);
-        outputView.printMoveResult(car);
-    }
-
-    private String getWinnerNames() {
-        Optional<Car> maxDistanceCar = carList.stream().max(Comparator.comparingInt(Car::getDistance));
-        Integer maxDistance = maxDistanceCar.get().getDistance();
-
-        return carList.stream().filter(car -> car.getDistance() == maxDistance)
-                .map(Car::getName)
-                .collect(Collectors.joining(", "));
-    }
-
-}
